# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CryoET Catalog API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/toml/{kind}": {
      "post": {
        "tags": [
          "toml"
        ],
        "summary": "Author Toml",
        "operationId": "author_toml_toml__kind__post",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/toml/{kind}/parse": {
      "post": {
        "tags": [
          "toml"
        ],
        "summary": "Parse Toml",
        "description": "Seed mode: upload (ADR-0004). Parse a posted ``.toml`` body into form\nstate with the backend ``tomllib`` loader \u2014 keeps TOML parsing off the\nfrontend. Unknown/extra keys come through verbatim; the renderer splits\nregistry fields from extras. Bad TOML -> 422 (no validation here; schema\nrules run on generate).",
        "operationId": "parse_toml_toml__kind__parse_post",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/toml/md-run-ids/{sample_id}": {
      "get": {
        "tags": [
          "toml"
        ],
        "summary": "Md Run Ids",
        "description": "md_source.md_run_id suggestions: the ids of the sample's known md_runs.\nFree-text remains accepted on submit; this only feeds the dropdown.",
        "operationId": "md_run_ids_toml_md_run_ids__sample_id__get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/toml/tilt-series-ids/{sample_id}/{acquisition_id}": {
      "get": {
        "tags": [
          "toml"
        ],
        "summary": "Tilt Series Ids",
        "description": "Tilt-series ids in one acquisition, for the reconstruction form.\n\nraw_tomogram.derived_from names a [[tilt_series]] in the acquisition's\nacquisition.toml \u2014 a different file, so the renderer's in-form cross-ref\npooling cannot supply it.",
        "operationId": "tilt_series_ids_toml_tilt_series_ids__sample_id___acquisition_id__get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          },
          {
            "name": "acquisition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acquisition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/toml/reconstruction-group-ids/{sample_id}/{acquisition_id}": {
      "get": {
        "tags": [
          "toml"
        ],
        "summary": "Reconstruction Group Ids",
        "description": "Alignment-group folder names in one acquisition, for the reconstruction\nform's group selector.\n\nOne reconstruction.toml per Reconstructions/<group>/ folder, so the form\nneeds the acquisition's group list to offer switching between them. The\nauthor route carries no sample context of its own, so it can't read the\nacquisition detail payload \u2014 hence this ids endpoint, mirroring\n/tilt-series-ids.",
        "operationId": "reconstruction_group_ids_toml_reconstruction_group_ids__sample_id___acquisition_id__get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          },
          {
            "name": "acquisition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Acquisition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/toml/{kind}/load/{record_id}": {
      "get": {
        "tags": [
          "toml"
        ],
        "summary": "Load Toml",
        "description": "Seed mode: pull-from-API (ADR-0004). Load an existing record's authored\nfields by id from the catalog DB. The data may lag the on-disk file \u2014 the\nrenderer surfaces a staleness warning for this mode.",
        "operationId": "load_toml_toml__kind__load__record_id__get",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "sample_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sample Id"
            }
          },
          {
            "name": "acquisition_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Acquisition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}